Safari-611.1.10/LayoutTests/http/tests/referrer-policy-script/strict-origin-when-cross-origin/cross-origin-http.https-expected.txt

Tests the behavior of strict-origin-when-cross-origin referrer policy when cross origin from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
